requirements.txt
Flask==0.11.1
Pillow==3.4.2
PyVirtualDisplay==0.2.1
selenium==3.4.3
